refactor(rocky-core): relocate TypedColumn / RockyType / LineageEdge from rocky-compiler
PR-1 of Typed-IR Option B Phase 1 (per ratified design memo at
rocky-plans/plans/typed-ir/option-b-design.md, Q3 ratification).

Move the typed-program primitives that ModelIr / ProjectIr will reference
into rocky-core, breaking the dependency cycle that PR-2 would otherwise
hit (rocky-compiler depends on rocky-core, one-way verified).

Relocations:
- rocky-core/src/types.rs (new): RockyType, StructField, TypedColumn,
  common_supertype, is_assignable + their tests.
- rocky-core/src/lineage.rs (new): QualifiedColumn, LineageEdge.
  TransformKind stays in rocky-sql (rocky-core already depends on rocky-sql).

Re-exports in rocky-compiler so existing call sites (rocky-ai, rocky-cli,
rocky-server, etc. — 12 external import sites) keep compiling unchanged:
- rocky-compiler/src/types.rs now re-exports rocky_core::types::*.
- rocky-compiler/src/semantic.rs re-exports LineageEdge / QualifiedColumn
  from rocky_core::lineage; SemanticGraph / ModelSchema / ColumnDef /
  build_semantic_graph stay in rocky-compiler (not in scope for this PR).

No behavior change. No public API change. No edits to rocky-cli/output.rs
or any downstream caller. Canonical-JSON convention call: relocated types
have no Option fields and no skip_serializing_if today, so the carry-over
is "no skip_serializing_if added" — Rule A vs B is deferred to PR-2 when
ModelIr / ProjectIr introduce actual Option fields. Doc comment on
rocky-core::traits::TypeMapper updated to point at the new RockyType
location.

Verification: cargo build --workspace clean, cargo test --workspace
--no-fail-fast all green, cargo clippy --workspace --all-targets -- -D
warnings clean, cargo fmt --check clean.
