+++ engine/crates/rocky-core/src/lib.rs
@@ -23,6 +23,7 @@ pub mod idempotency;
 pub mod incremental;
 pub mod ir;
 pub mod lakehouse;
+pub mod lineage;
 pub mod macros;
 pub mod masking;
 pub mod mmap;
@@ -48,5 +49,6 @@ pub mod state;
 pub mod state_sync;
 pub mod tests;
 pub mod traits;
+pub mod types;
 pub mod unified_dag;
 pub mod unit_test;
